Only parse argv for commands
The bot currently tries to parse an argv array for each channel message
it receives. This results in it sending unterminated quote messages in
response to any chat message which contains quotes.
Since argv arrays are only necessary for bot commands, a change has been
made so that it only parses a chat message into an array if the chat
message begins with the command prefix.